[NVMESH-8479] fix (Toma): shrink topo_idx_updated to 32 bits
Every consumer truncates to the lower 32 bits via extract_lower_32_bits_idx() or compares globally-monotonic lifecycle counter values, so the packed term in the upper bits is dead weight. Drop it and restore the freed wire slot to a real reserved field. Wire layout stays byte-identical.

+++ toma/unitest/00_framework/toma_test_helpers.h
@@ -19,7 +19,7 @@ extern void TEST_set_nm_remote_nodes_connected(bool is_connected);
 
 /* Praid hash — nvmeibt_praid.c */
 extern void TEST_init_praids_hash(void);
-extern void TEST_add_praid_to_hash(const union nvmeib_uuid *uuid, int64_t topo_idx_updated,
+extern void TEST_add_praid_to_hash(const union nvmeib_uuid *uuid, int32_t topo_idx_updated,
 								   int praid_version_major, int praid_version_minor);
 extern void TEST_set_praid_conf_corrupted(const union nvmeib_uuid *uuid);
